requirements.txt
Flask==2.2.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
numpy==1.23.2
pandas==1.4.4
sklearn
scikit-learn==1.1.2
SQLAlchemy==1.4.41
xgboost==1.6.1
gunicorn==20.0.4
python-dotenv
psycopg2